Fix precision issues in token transfers and encryption
- Use BigInt for SPL token transfer amounts to prevent overflow
- Fix encryption to avoid data corruption from hex conversion
- Add proper token decimal handling in balance checks
- Improve terminology: 'Dev fee' -> 'Development fee'
- Add comments about token amount precision requirements

+++ src/services/profitDistribution.ts
@@ -112,12 +112,14 @@ export class ProfitDistributionService {
         );
 
         // Add transfer instructions
+        // Note: Token amounts should already be in the smallest unit (e.g., lamports for SOL)
+        // If totalProfit is in token decimals, it needs to be converted first
         transaction.add(
           createTransferInstruction(
             payerTokenAccount,
             reserveTokenAccount,
             payerKeypair.publicKey,
-            Math.floor(reserveAmount)
+            BigInt(Math.floor(reserveAmount))
           )
         );
 
@@ -126,7 +128,7 @@ export class ProfitDistributionService {
             payerTokenAccount,
             gasTokenAccount,
             payerKeypair.publicKey,
-            Math.floor(gasAmount)
+            BigInt(Math.floor(gasAmount))
           )
         );
 
@@ -135,7 +137,7 @@ export class ProfitDistributionService {
             payerTokenAccount,
             daoTokenAccount,
             payerKeypair.publicKey,
-            Math.floor(daoAmount)
+            BigInt(Math.floor(daoAmount))
           )
         );
       } else {
